puts all components on a single page with teamSetup being moved to the left
