+++ src/components/NumTeams.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import '../App.css';
-import { Button, TextField, InputLabel, Select, MenuItem, FormControl } from '@material-ui/core';
+import { InputLabel, Select, MenuItem, FormControl } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
-import { Link } from 'react-router-dom';
 
 const useStyles = makeStyles(theme => ({
   button: {
@@ -22,7 +21,7 @@ const useStyles = makeStyles(theme => ({
   }
 }));
 
-const NumTeams = (props) => {
+const NumTeams = ({ setNumTeams }) => {
 
     const classes = useStyles();
 
@@ -30,12 +29,11 @@ const NumTeams = (props) => {
         <div>
           <FormControl className={classes.formControl}>
             <InputLabel id="label-id">Number of Teams</InputLabel>
-            <Select labelId="label-id" id="select" onChange={props.setNumTeams} className={classes.select}>
+            <Select labelId="label-id" id="select" onChange={setNumTeams} className={classes.select}>
               {Array(16).fill(1).map((num, i) => (
                   <MenuItem key={i + 1} value={i + 1}>{i + 1}</MenuItem>
               ))}
             </Select>
-            {/* <Link to='/teams'><Button variant="contained" color="primary">Submit</Button></Link> */}
           </FormControl>
         </div>
     )
