+++ src/components/TeamSetupForm.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useState } from 'react';
-import { TextField, Input } from '@material-ui/core';
+import { TextField } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
